product/[encodedUrl]->/api/product/[id]
