+++ app/product/[id]/layout.tsx
@@ -4,8 +4,8 @@ import React from "react";
 import type { Metadata } from "next";
 import { getProduct } from "@/lib/component-actions";
 
-export async function generateMetadata({ params }: { params: { encodedUrl: string } }): Promise<Metadata> {
-  const product = await getProduct(params.encodedUrl).catch((error) => {
+export async function generateMetadata({ params }: { params: { id: string } }): Promise<Metadata> {
+  const product = await getProduct(params.id).catch((error) => {
     console.error("[ProductLayout]:", error);
     return null;
   });
